+++ lib/osf-components/addon/components/contributor-list/component.ts
@@ -3,59 +3,79 @@ import { computed } from '@ember-decorators/object';
 import { alias } from '@ember-decorators/object/computed';
 import { service } from '@ember-decorators/service';
 import Component from '@ember/component';
+import { task } from 'ember-concurrency';
 import DS from 'ember-data';
 import I18N from 'ember-i18n/services/i18n';
 
 import { layout } from 'ember-osf-web/decorators/component';
 import Contributor from 'ember-osf-web/models/contributor';
+import Node from 'ember-osf-web/models/node';
+import { QueryHasManyResult } from 'ember-osf-web/models/osf-model';
+import Ready from 'ember-osf-web/services/ready';
 import defaultTo from 'ember-osf-web/utils/default-to';
+
+import styles from './styles';
 import template from './template';
 
-export interface Contrib {
-    title: string;
-    id: string | null;
-}
-
-@layout(template)
+@layout(template, styles)
 @tagName('span')
-export default class ContributorList extends Component {
+export default class ContributorList extends Component.extend({
+    didReceiveAttrs(this: ContributorList) {
+        this.loadContributors.perform();
+    },
+
+    loadContributors: task(function *(this: ContributorList, more?: boolean) {
+        if (!this.node) {
+            return;
+        }
+
+        const blocker = this.ready.getBlocker();
+        if (this.shouldLoadAll && !this.shouldTruncate) {
+            const allContributors = yield this.node.loadAll('contributors');
+            this.setProperties({
+                displayedContributors: allContributors.toArray(),
+                totalContributors: allContributors.length,
+            });
+        } else if (more) {
+            const nextPage: QueryHasManyResult<Contributor> = yield this.node.queryHasMany('contributors', {
+                page: this.incrementProperty('page'),
+            });
+            this.displayedContributors.pushObjects(nextPage);
+            this.set('totalContributors', nextPage.meta.total);
+        } else {
+            this.set('page', 1);
+            const firstPage = yield this.node.contributors;
+            this.setProperties({
+                displayedContributors: firstPage.toArray(),
+                totalContributors: firstPage.meta.total,
+            });
+        }
+
+        blocker.done();
+    }).drop(),
+}) {
+    // Required arguments
+    node?: Node;
+
+    // Optional arguments
+    shouldTruncate: boolean = defaultTo(this.shouldTruncate, true);
+    shouldLinkUsers: boolean = defaultTo(this.shouldLinkUsers, false);
+
+    // Private properties
     @service i18n!: I18N;
-    max = 3;
-    truncated: boolean = defaultTo(this.truncated, true);
-    nodeId?: string;
-    useLink?: boolean;
-    contributors!: DS.PromiseManyArray<Contributor> & { meta: { total: number } };
+    @service store!: DS.Store;
+    @service ready!: Ready;
 
-    @computed('contributors.[]')
-    get contributorList(this: ContributorList): Contrib[] {
-        if (!this.contributors) {
-            return [];
-        }
-        const contributors = this.contributors.toArray();
+    page = 1;
+    displayedContributors: Contributor[] = [];
+    totalContributors?: number;
+    shouldLoadAll: boolean = navigator.userAgent.includes('Prerender');
 
-        if (!(contributors && contributors.length)) {
-            return [];
-        }
-
-        const contribs: Contrib[] = contributors
-            .map(c => ({
-                title: c.users.get('familyName') || c.users.get('givenName') || c.users.get('fullName'),
-                id: c.get('unregisteredContributor') ? null : c.users.get('id'),
-            }));
-
-        return contribs;
-    }
-
-    @alias('contributorList.firstObject') first!: Contrib;
-    @alias('contributors.meta.total') numContributors!: number;
-
-    @computed('contributors.meta.total', 'max')
-    get rest(this: ContributorList): number {
-        return this.contributors.meta.total - this.max;
-    }
+    @alias('loadContributors.isRunning')
+    isLoading!: boolean;
 
     @computed('truncated')
-    get truncate() {
-        return this.truncated ? 3 : undefined;
+    get truncateCount() {
+        return this.shouldTruncate ? 3 : undefined;
     }
 }