+++ lib/osf-components/addon/components/panel/component.ts
@@ -1,8 +1,10 @@
-import { layout } from '@ember-decorators/component';
 import Component from '@ember/component';
 
+import { layout } from 'ember-osf-web/decorators/component';
+
+import styles from './styles';
 import template from './template';
 
-@layout(template)
+@layout(template, styles)
 export default class Panel extends Component {
 }