+++ mirage/factories/user.ts
@@ -18,13 +18,14 @@ export default Factory.extend<User & UserTraits>({
     afterCreate(user: ModelInstance<User>, server: Server) {
         guidAfterCreate(user, server);
         server.create('user-email', { user, primary: true });
+
         const defaultRegion = server.schema.regions.find('us');
         user.update({ defaultRegion });
+        if (!user.fullName && (user.givenName || user.familyName)) {
+            user.update('fullName', [user.givenName, user.familyName].filter(Boolean).join(' '));
+        }
     },
 
-    fullName() {
-        return `${faker.name.firstName()} ${faker.name.lastName()}`;
-    },
     givenName() {
         return faker.name.firstName();
     },