+++ tests/engines/registries/acceptance/overview/comments-test.ts
@@ -11,10 +11,6 @@ module('Registries | Acceptance | overview.comments', hooks => {
     setupEngineApplicationTest(hooks, 'registries');
     setupMirage(hooks);
 
-    hooks.beforeEach(function(this: TestContext) {
-        server.create('user', 'loggedIn');
-    });
-
     test('it renders', async function(this: TestContext, assert: Assert) {
         const registration = server.create(
             'registration',