+++ tests/unit/mirage/views/utils/sort-test.ts
@@ -7,7 +7,7 @@ import {
     dynamicSort,
     ProcessOptions,
     sort,
-} from 'ember-osf-web/mirage/views/private/utils';
+} from 'ember-osf-web/mirage/views/utils/-private';
 
 function fakeRequest(queryParams: { [key: string]: string }): Request {
     return {