+++ CHANGELOG.md
@@ -20,8 +20,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `developer-app` - extend `links` types to include links specific to developer apps
     - `collection` - improve types for choices fields
     - `collected-metadatum` - improve types for choice fields
+- Components:
+    - `<ContributorList>` - modified to take in different parameters
+        - `node` (required)
+        - `shouldTruncate` (default true)
+        - `shouldLinkUsers` (default false)
 
-## [19.1.0]
+## [19.1.0] - 2019-01-23
 ### Added
 - Addons:
     - `loaders-css` - For animated loading indicators
@@ -34,6 +39,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `settings.account`
 - Styles:
     - All styles from `osf-style`
+- Validators:
+    - `httpUrl` - validates that a string looks like an http url
 
 ### Changed
 - Addons:
@@ -43,6 +50,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `ember-cli-sass@8.0.1`
 - Components:
     - `contributor-list` - takes an optional parameter `truncated`
+- Models:
+    - `developer-app` - use custom `httpUrl` validator for urls
 - Templates:
     - `no-implicit-this` template rule activated
 - Types: